# openapi/honest-license-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HonestLicenseServer",
    "version": "1.0"
  },
  "paths": {
...
    "/api/assets/{component}/{version}/download": {
      "get": {
        "tags": [
          "Assets"
        ],
        "operationId": "Assets_Download",
        "parameters": [
          {
            "name": "component",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/assets/install/{component}/{version}/download": {
      "get": {
        "tags": [ "Assets" ],
        "operationId": "Assets_InstallDownload",
        "parameters": [
          { "name": "component", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "version", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "302": { "description": "Found" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" },
          "502": { "description": "Bad Gateway" }
        },
        "security": [ { "InstallBearer": [ ] } ]
      }
    },
    "/api/service/install-access": {
      "post": {
        "tags": [ "ServiceInstallation" ],
        "operationId": "ServiceInstallation_Authorize",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ServiceInstallAccessRequest" } },
            "text/json": { "schema": { "$ref": "#/components/schemas/ServiceInstallAccessRequest" } },
            "application/*+json": { "schema": { "$ref": "#/components/schemas/ServiceInstallAccessRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServiceInstallAccessResponse" } }
            }
          },
          "401": { "description": "Invalid service installation password" },
          "403": { "description": "Service installation access is disabled" },
          "429": { "description": "Too Many Requests" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ServiceInstallAccessRequest": {
        "required": [ "password" ],
        "type": "object",
        "properties": {
          "password": { "type": "string", "minLength": 1, "maxLength": 512 },
          "appVersion": { "type": "string", "maxLength": 64, "nullable": true },
          "architecture": { "type": "string", "maxLength": 20, "nullable": true }
        },
        "additionalProperties": false
      },
      "ServiceInstallAccessResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "expiresInSeconds": { "type": "integer", "format": "int32" },
          "scope": { "type": "string" },
          "components": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ServiceInstallComponentResponse" }
          }
        },
        "additionalProperties": false
      },
...
      "ServiceInstallComponentResponse": {
        "type": "object",
        "properties": {
          "component": { "type": "string" },
          "version": { "type": "string" },
          "fileName": { "type": "string", "nullable": true },
          "downloadUrl": { "type": "string", "nullable": true },
          "sha256": { "type": "string", "nullable": true },
          "sizeBytes": { "type": "integer", "format": "int64", "nullable": true },
          "architecture": { "type": "string", "nullable": true }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token returned by POST /api/auth/login or /api/auth/refresh.",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "AdminKey": {
        "type": "apiKey",
        "description": "Administrative API key used by HonestDesk.",
        "name": "X-Admin-Key",
        "in": "header"
      },
      "InstallBearer": {
        "type": "http",
        "description": "Short-lived installation_only token returned by POST /api/service/install-access.",
        "scheme": "bearer",
        "bearerFormat": "opaque-install-only"
      }
    }
  },
...
}